remove debian packaging, release individual files
Debian packages are created [here]( https://github.com/Redecorating/mbp-ubuntu-kernel/releases), and someone said that works, so that is no longer needed.

Previously on the releases page there'd be a zip file with the packages, but now it should do seperate files as a different action is now being used.